Revert ":recycle: refactor: Adicionando novos gráficos Dashboard"
