+++ src/components/Dashboard/GraficoDebito.jsx
@@ -1,20 +1,53 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Bar } from "react-chartjs-2";
+import createAPI from "../../services/createAPI";
 
-const GraficoDebito = ({ dados = [], totalAtivos = 0 }) => {
-  const hasData = dados && Array.isArray(dados) && dados.length > 0;
-  
+const GraficoDebito = () => {
+  const [dados, setDados] = useState([]);
+  const [totalAtivos, setTotalAtivos] = useState(0);
+
+   useEffect(() => {
+    const debito = async () => {
+      const requisicao = createAPI();
+      try {
+        const response = await requisicao.get("/veiculo/debito/cont");
+        if (response.status === 200 && response.data.msg.resultado) {
+          const newData = response.data.data;
+          setDados(newData);
+          setTotalAtivos(newData[newData.length - 1].total_debitos_ativos);
+
+        } else {
+          setDados([]);
+          setTotalAtivos(0);}
+      } catch (error) {
+        if (
+          error?.response?.data?.msg === "Cabeçalho inválido!" ||
+          error?.response?.data?.msg === "Token inválido!" ||
+          error?.response?.data?.msg ===
+            "Usuário não possui o perfil mencionado!"
+        ) {
+          localStorage.removeItem("user");
+          localStorage.removeItem("token");
+          localStorage.removeItem("perfil");
+        } else {
+          console.log(error);
+        }
+      }
+    };
+    debito();
+  }, []);
+
   const chartData = {
-    labels: hasData ? dados.map(item => item.mes) : [],
+    labels: dados.map(item => item.mes),
     datasets: [
       {
         label: "Ativações",
-        data: hasData ? dados.map(item => Number(item.debitos_ativados || 0)) : [],
+        data: dados.map(item => Number(item.debitos_ativados)),
         backgroundColor: "#4594E5",
       },
       {
         label: "Desativações",
-        data: hasData ? dados.map(item => Number(item.debitos_desativados || 0)) : [],
+        data: dados.map(item => Number(item.debitos_desativados)),
         backgroundColor: "#FA5F5F",
       },
     ],
@@ -24,10 +57,6 @@ const GraficoDebito = ({ dados = [], totalAtivos = 0 }) => {
     responsive: true,
     maintainAspectRatio: false,
     indexAxis: 'x',
-    interaction: {
-      mode: 'index',
-      intersect: false,
-    },
     plugins: { 
       legend: { 
         position: "top",
@@ -45,7 +74,6 @@ const GraficoDebito = ({ dados = [], totalAtivos = 0 }) => {
         }
       },
       y: {
-        beginAtZero: true,
         ticks: {
           autoSkip: false 
         }
@@ -59,18 +87,12 @@ const GraficoDebito = ({ dados = [], totalAtivos = 0 }) => {
         Débitos Ativos <span style={{ color: "#4594E5" }}>{totalAtivos}</span>
       </h4>
       <div style={{ width: "100%", height: "340px" }}>
-        {hasData ? (
-          <Bar 
-            data={chartData} 
-            options={options}
-            height={null}
-            width={null} 
-          />
-        ) : (
-          <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '100%', color: '#888' }}>
-            Carregando dados...
-          </div>
-        )}
+        <Bar 
+          data={chartData} 
+          options={options}
+          height={null}
+          width={null} 
+        />
       </div>
     </div>
   );
